# Optimized the time to save the asset loss table