--- a/debian/changelog
+++ b/debian/changelog
@@ -1,4 +1,5 @@
   [Michele Simionato]
+  * Optimized the saving of the asset loss table
   * Fixed HDF5 segmentation faults in the tests for Ubuntu 16.04
 
   [Daniele Viganò]