I tested this on the time-honored Germany computation. With this optimization there is a speedup of 800 times in the saving time, from 40 minutes to 3 seconds; without optization the calculation is stuck waiting for the asset loss table to be written in the datastore, even it is faily small, only 103.42 MB. Here are the numbers on Wilson:
```
   with optimization
   ======================== ========= ========= =========
   operation                time_sec  memory_mb counts   
   ======================== ========= ========= =========
   total event_based_risk   85,293    1,465     71       
   computing riskmodel      44,637    0.0       1,075,721
   building hazard          37,526    2,890     71       
   compute poes             37,303    0.0       31,553   
   aggregate losses         32,694    0.0       1,075,721
   saving event loss tables 2.948     22        71       
   ======================== ========= ========= =========
  data transfer back: 104 MB

  without optimization
   ============================== ========= ========= =========
   operation                      time_sec  memory_mb counts   
   ============================== ========= ========= =========
   total event_based_risk         98,227    1,532     71       
   computing riskmodel            57,367    0.0       1,075,721
   building hazard                37,825    2,812     71       
   compute poes                   37,576    0.0       31,553   
   aggregate losses               32,138    0.0       1,075,721
   saving event loss tables       2,387     174       71       
   ============================== ========= ========= =========
  data transfer back: 683 MB
``` 
The tests are running here: https://ci.openquake.org/job/zdevel_oq-risklib/1536/